Hack: set path via RG_PATH env var
It would be preferable to set the path as $2 but I've struggled to
make that change without breaking handling of the regex pattern in $1.
Specifically, attempts to do so broke the use of \( to match ( and
appeared also to break the interpretation of regex special characters
such as . What does work is using "$*" to set the regex pattern; but I
haven't yet found a way of removing $2 from that without breaking things.